Convert string pool to an array of lstring structures, lots of low-level changes.
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3217 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/tex/maincontrol.c
@@ -3190,7 +3190,7 @@ void issue_message(void)
     s = make_string();
     if (c == 0) {
         /* Print string |s| on the terminal */
-        if (term_offset + str_length(s) > max_print_line - 2)
+        if (term_offset + (int)str_length(s) > max_print_line - 2)
             print_ln();
         else if ((term_offset > 0) || (file_offset > 0))
             print_char(' ');
